Added Decimals to handle bigger numbers

+++ factorial/factorial.py
@@ -1,5 +1,6 @@
 from subprocess import run, call, PIPE
 from pathlib import Path
+from decimal import Decimal
 
 
 def factorial(n):
@@ -20,4 +21,4 @@ def factorial(n):
     result = run([f'{Path(__file__).parent.resolve() / "factorial"}', str(n)], stdout=PIPE, encoding='utf-8')
 
     # I have to go via float to read scientific notation to int
-    return int(float(result.stdout.strip()))
+    return int(Decimal(result.stdout.strip()))
